+++ src/order.cpp
@@ -7,7 +7,7 @@ Order::Order(
     Side side, 
     OrderType type, 
     int quantity, 
-    double price, 
+    long long price, 
     long long timestamp
 )
 {
