+++ src/trade.cpp
@@ -7,7 +7,7 @@ Trade::Trade(
     long long buyOrderId, 
     long long sellOrderId, 
     int quantity, 
-    double price, 
+    long long price, 
     long long timestamp
 )
 {
